Update Mesh class
